Temporary Shot Glass Image
First version of shot glass texture. Feedback is needed on card size
(The cream colour, whether or not rounded corners are wanted on the
card) and the shot glass design (Straight edges/more angular like the
logo). All feedback on this image must be made before any other glass
images are made.